Handling opposite set in brackets with ^
